requirements.txt
torch
torchvision
pillow
pandas
yaml
tensorboardX
pytz